Changed my mind, doing validation in the viewmodel.
Easier to do, more separation between program logic and view, and i
can actually figure out how to disable the view/send transaction
button command when any of the pending outputs are invalid.

This validates the address/script field but amounts are still using
the validation class. need to switch that over next...